Rewrote all uses of Song to Alchemy and fiddled a few things in object bases. Resolves #29.
Lots of zombies now exist in the code files - songs still exist, they just depend on alchemy and can't be acquired. This is horrifying. But I'm afraid to take them out - what if I want a song-of-silence-like-effect but deleted all the song of silence code?! I can backcopy from narsil, but I might miss some of the pieces (unlikely).