Simplify logic when tunnel's start is same as the end
Makes comparisons with Sil 1.3 easier (no special handling of such tunnels there) and removes the asymmetric handling of the vertical and horizontal cases (only had special treatment of such tunnels in the horizontal case).  Does not appear to affect tunnel generation:  such tunnels are either prevented by the room layout or are eliminated by the other tests in tunnel_ok().